Split the progect into {program, lib, bench}, remove weird CR
