+++ README.md
@@ -1,5 +1,15 @@
-# To start the spell checker using cabal
-cabal run
+# Spell Checker
 
-# To stop it
+## To start the spell checker using cabal
+``` sh
+cabal run spell-checker
+```
+
+## To run the bench
+``` sh
+cabal bench
+```
+
+
+## To stop it
 Type enter
