+++ bench/Bench.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Data.WordTree
+
+main :: IO ()
+main = putStrLn "Hello Bench"
