+++ spell-checker.cabal
@@ -1,6 +1,7 @@
 cabal-version:      3.6
 name:               spell-checker
 version:            0.1.0.0
+build-type: Simple
 
 -- A short (one-line) description of the package.
 -- synopsis: Spell-checker in Haskell
@@ -15,7 +16,7 @@ version:            0.1.0.0
 -- license: aucune
 
 -- The package author(s).
--- author: Nathan Maillet
+author: Nathan Maillet
 
 -- An email address to which users can send suggestions, bug reports, and patches.
 -- maintainer:
@@ -25,14 +26,10 @@ version:            0.1.0.0
 -- category:
 extra-source-files: CHANGELOG.md
 
-executable spell-checker
-    main-is:    Main.hs
+library spell-checker-lib
+    exposed-modules:
+        Data.WordTree
 
-    -- Modules included in this executable, other than Main.
-    other-modules: WordsTrees, StatsFromTxt
-
-    -- LANGUAGE extensions used by modules in this package.
-    -- other-extensions:
     build-depends:
         base        ^>= 4.16.3.0,
         parallel    ^>= 1.0.0.0,
@@ -42,6 +39,46 @@ executable spell-checker
         containers  ^>= 0.6.0.0,
         directory   ^>= 1.3.0.0,
         mtl         ^>= 2.3.1
-        --text ^>=1.2.4.1
+
+    hs-source-dirs: lib
+    default-language: Haskell2010
+
+executable spell-checker
+    main-is:    Main.hs
+
+    -- Modules included in this executable, other than Main.
+    other-modules: StatsFromTxt
+
+    -- LANGUAGE extensions used by modules in this package.
+    -- other-extensions:
+    build-depends:
+        spell-checker:spell-checker-lib,
+        base        ^>= 4.16.3.0,
+        parallel    ^>= 1.0.0.0,
+        vector      ^>= 0.12.0.0,
+        aeson       ^>= 2.1.0.0,
+        bytestring  ^>= 0.11.3.1,
+        containers  ^>= 0.6.0.0,
+        directory   ^>= 1.3.0.0,
+        mtl         ^>= 2.3.1
     hs-source-dirs:   app
+    default-language: Haskell2010
+
+
+benchmark spell-checker-bench
+    type: exitcode-stdio-1.0
+    main-is: Bench.hs
+    build-depends:
+        spell-checker:spell-checker-lib,
+        base        ^>= 4.16.3.0,
+        parallel    ^>= 1.0.0.0,
+        vector      ^>= 0.12.0.0,
+        aeson       ^>= 2.1.0.0,
+        bytestring  ^>= 0.11.3.1,
+        containers  ^>= 0.6.0.0,
+        directory   ^>= 1.3.0.0,
+        mtl         ^>= 2.3.1
+
+
+    hs-source-dirs: bench
     default-language: Haskell2010
